# api_schemas/openapi.yaml
...
openapi: 3.0.3
info:
  title: Jarvis AI API v1
  description: |
    Local-first AI orchestration platform providing chat, persona management, monitoring, and system configuration capabilities.
    
    ## Authentication
    All endpoints require authentication via the `X-API-Key` header. Set `JARVIS_DISABLE_AUTH=true` to bypass authentication locally.
    
    ## Endpoint Categories
    - **Core API** (`/api/v1/*`): Main Jarvis functionality
    - **Management API** (`/api/v1/management/*`): System configuration and persona management
    - **OpenAI-Compatible** (`/v1/*`): Drop-in replacement for OpenAI API clients
  version: "1.0.0"
  contact:
    name: Jarvis AI Support
    email: support@jarvis.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8000
    description: Local development server
  - url: https://api.jarvis.ai
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/chat:
    post:
      summary: Chat completion
      description: Generate a chat response using the specified persona and AI models
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              messages:
                - role: user
                  content: "Hello, how are you?"
              persona: "generalist"
              temperature: 0.7
              max_tokens: 512
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              example:
                content: "I'm doing well, thank you for asking! How can I help you today?"
                model: "ollama"
                tokens: 25
                diagnostics: {}
        '400':
          description: Bad request (e.g., persona not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # Core API - Monitoring
  # ============================================================
...
  /v1/chat/completions:
    post:
      summary: OpenAI-compatible chat completions
      description: |
        Drop-in replacement for OpenAI's chat completions API.
        The `model` parameter is mapped to a Jarvis persona if it exists, otherwise defaults to "generalist".
      tags: [OpenAI Compatible]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAIChatRequest'
            example:
              model: "jarvis-default"
              messages:
                - role: user
                  content: "Hello!"
              temperature: 0.7
              max_tokens: 512
              stream: false
      responses:
        '200':
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIChatResponse'
              example:
                id: "chatcmpl-1702648475"
                object: "chat.completion"
                created: 1702648475
                model: "ollama"
                choices:
                  - index: 0
                    message:
                      role: "assistant"
                      content: "Hello! How can I help you today?"
                    finish_reason: "stop"
                usage:
                  prompt_tokens: 10
                  completion_tokens: 15
                  total_tokens: 25
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail

    # ============================================================
    # Health
    # ============================================================
...
    Message:
      type: object
      properties:
        role:
          type: string
          description: Role of the speaker (user, assistant, system)
        content:
          type: string
          description: Message content
      required:
        - role
        - content
    ChatRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: Conversation messages
        persona:
          type: string
          default: "generalist"
          description: Persona to route the request
        temperature:
          type: number
          format: float
          default: 0.7
          minimum: 0.0
          maximum: 2.0
          description: Controls randomness in generation
        max_tokens:
          type: integer
          default: 512
          minimum: 32
          maximum: 4096
          description: Maximum tokens to generate
        metadata:
          type: object
          description: Optional metadata for the request
        external_context:
          type: array
          items:
            type: string
          description: Optional external context strings
      required:
        - messages
    ChatResponse:
      type: object
      properties:
        content:
          type: string
          description: Generated response text
        model:
          type: string
          description: Backend model used for generation
        tokens:
          type: integer
          description: Number of tokens generated
        diagnostics:
          type: object
          description: Backend-specific diagnostic information
      required:
        - content
        - model
        - tokens
        - diagnostics

    # ============================================================
    # Persona Models
    # ============================================================
...
    OpenAIChatRequest:
      type: object
      properties:
        model:
          type: string
          default: "jarvis-default"
          description: Model name (mapped to persona if exists)
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: Conversation messages
        temperature:
          type: number
          format: float
          default: 0.7
          description: Controls randomness
        max_tokens:
          type: integer
          default: 512
          minimum: 32
          maximum: 4096
          description: Maximum tokens to generate
        stream:
          type: boolean
          default: false
          description: Whether to stream response (not yet implemented)
      required:
        - messages
    OpenAIChatResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique completion ID
        object:
          type: string
          default: "chat.completion"
        created:
          type: integer
          description: Unix timestamp
        model:
          type: string
          description: Model used
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
      required:
        - id
        - object
        - created
        - model
        - choices
        - usage
...